absorb geometry, timer sources into header, remove src folder, always enable SORT_BACK
