+++ include/build_tree.h
@@ -71,9 +71,7 @@ namespace exafmm_t {
       int octant = (x[0] > X[0]) + ((x[1] > X[1]) << 1) + ((x[2] > X[2]) << 2);
       buffer[counter[octant]].X = bodies[i].X;
       buffer[counter[octant]].q = bodies[i].q;
-#if SORT_BACK
       buffer[counter[octant]].ibody = bodies[i].ibody;
-#endif
       counter[octant]++;
     }
   }
@@ -125,18 +123,14 @@ namespace exafmm_t {
           for (int d=0; d<3; ++d) {
             node->src_coord.push_back(B->X[d]);
           }
-#if SORT_BACK
           node->isrcs.push_back(B->ibody);
-#endif
           node->src_value.push_back(B->q);
         }
         for (Body<T>* B=first_target; B<first_target+node->ntrgs; ++B) {
           for (int d=0; d<3; ++d) {
             node->trg_coord.push_back(B->X[d]);
           }
-#if SORT_BACK
           node->itrgs.push_back(B->ibody);
-#endif
         }
       }
       return;
