+++ include/timer.h
@@ -12,16 +12,17 @@ namespace exafmm_t {
   static const int decimal = 7;                 //!< Decimal precision
   static const int wait = 100;                  //!< Waiting time between output of different ranks
   static const int dividerLength = stringLength + decimal + 9;  // length of output section divider
-  extern long long flop;
-  extern timeval time;
-  extern std::map<std::string, timeval> timer;
+  long long flop = 0;
+  timeval time;
+  std::map<std::string, timeval> timer;
 
-  void start(std::string event);
-  double stop(std::string event, bool verbose=true);
-  void print(std::string s);
-  void print_divider(std::string s);
-
-  void add_flop(long long n);
+  void print(std::string s) {
+    // if (!VERBOSE | (MPIRANK != 0)) return;
+    s += " ";
+    std::cout << "--- " << std::setw(stringLength) << std::left
+              << std::setfill('-') << s << std::setw(decimal+1) << "-"
+              << std::setfill(' ') << std::endl;
+  }
 
   template<typename T>
   void print(std::string s, T v, bool fixed=true) {
@@ -33,6 +34,31 @@ namespace exafmm_t {
     std::cout << v << std::endl;
   }
 
+  void print_divider(std::string s) {
+    s.insert(0, " ");
+    s.append(" ");
+    int halfLength = (dividerLength - s.length()) / 2;
+    std::cout << std::string(halfLength, '-') << s
+              << std::string(dividerLength-halfLength-s.length(), '-') << std::endl;
+  }
 
+  void add_flop(long long n) {
+#pragma omp atomic update
+    flop += n;
+  }
+
+  void start(std::string event) {
+    gettimeofday(&time, NULL);
+    timer[event] = time;
+  }
+
+  double stop(std::string event, bool verbose=true) {
+    gettimeofday(&time, NULL);
+    double eventTime = time.tv_sec - timer[event].tv_sec +
+      (time.tv_usec - timer[event].tv_usec) * 1e-6;
+    if (verbose)
+      print(event, eventTime);
+    return eventTime;
+  }
 }
 #endif
